requirements.txt
Pillow==10.1.0
pyautogui==0.9.54
keyboard==0.13.5
numpy==1.24.3
PyAudio==0.2.13
av==16.0.1